SlideGeneratorAPI/requirements.txt
flask==2.3.3
python-pptx==0.6.21
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0